The engine now compiles and starts up

+++ Volt/Volt/src/Volt/Asset/Importers/NetContractImporter.cpp
@@ -57,7 +57,7 @@ namespace Volt
 				{
 					auto outerKey = it->first.as<entt::entity>();
 					std::unordered_map<std::string, NetRule> outerValue;
-					auto ruleNodeInner = ruleNodeOuter[std::to_string(outerKey).c_str()];
+					auto ruleNodeInner = ruleNodeOuter[std::to_string(static_cast<uint32_t>(outerKey)).c_str()];
 
 					for (YAML::const_iterator rulesIt = ruleNodeInner.begin(); rulesIt != ruleNodeInner.end(); ++rulesIt)
 					{
